+++ Assets/Script/Service/View/Game/Controller/PlayerController.cs
@@ -9,12 +9,11 @@ namespace Script.Service.View.Game.Controller
     {
         public override int Hp { get; set; }
 
-        public override int Speed { get; set; }
+        public override float Speed { get; set; }
 
         public override void Harm(HarmData data)
         {
         }
-
         private int _speed;
         private int _hp;
         private Rigidbody2D _rigidbody2D;
